v1.3
Fixed first load most of the time not actually spawning a grass block, resulting in a block with missing textures..
Changed the way drops are spawned to work with the new way drops behave when inside a block.
Changed how the check for an already-existing OneBlock world works. Now it just goes though all existing worlds, and if there is not one named "OneBlock", then it assumes one does not exist.